Split up ruby code

+++ compose_base/gen_lamps.rb
@@ -1,4 +1,5 @@
 require 'fileutils'
+require_relative '../compose_context'
 
 colors = [
   "white",
@@ -19,20 +20,34 @@ colors = [
 ]
 
 sizes = ["small", "large"]
-target_dirname = File.join(__dir__, "../compose_src/blocks/lamp")
 
-FileUtils.rm_rf target_dirname
-FileUtils.mkdir_p target_dirname
+ctx = Compose::Context.new('compose_base/gen_lamps')
+ctx.add_reference(nil, __FILE__)
 
 Dir.glob(File.join(__dir__, "lamp", "*.json")) do |filename|
-  contents = File.read(filename)
-  basename = File.basename(filename)
-  sizes.each do |size|
-    colors.each do |color|
-      new_contents = contents.gsub("$color$", color).gsub("$size$", size)
-      target_filename = File.join(target_dirname, "#{size}_#{color}_" + basename)
-      puts "WRITE: #{target_filename}"
-      File.write(target_filename, new_contents)
+  ctx.add_reference(nil, filename)
+end
+
+if ctx.modified
+  ctx.save_file()
+  puts "Lamps have changed"
+  target_dirname = File.join(__dir__, "../compose_src/blocks/lamp")
+
+  FileUtils.rm_rf target_dirname
+  FileUtils.mkdir_p target_dirname
+
+  Dir.glob(File.join(__dir__, "lamp", "*.json")) do |filename|
+    contents = File.read(filename)
+    basename = File.basename(filename)
+    sizes.each do |size|
+      colors.each do |color|
+        new_contents = contents.gsub("$color$", color).gsub("$size$", size)
+        target_filename = File.join(target_dirname, "#{size}_#{color}_" + basename)
+        puts "WRITE: #{target_filename}"
+        File.write(target_filename, new_contents)
+      end
     end
   end
+else
+  puts "Nothing to do."
 end
